requirements.txt
pandas
openpyxl
python-docx
requests
schedule
